updated the code based on review
