+++ client/src/components/Profile/Availabilty.tsx
@@ -1,7 +1,6 @@
 import Button from '@mui/material/Button';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
-import { Link } from 'react-router-dom';
 
 const Availabilty = (): JSX.Element => {
   return (
@@ -13,7 +12,7 @@ const Availabilty = (): JSX.Element => {
       alignSelf="flex-end"
       marginRight={5}
     >
-      <h1>Availabilty</h1>
+      <Typography style={{ fontSize: '30px', fontWeight: '100' }}>Availabilty</Typography>
     </Box>
   );
 };
