+++ client/src/components/Profile/EditProfile.tsx
@@ -1,21 +1,20 @@
 import Button from '@mui/material/Button';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
-import { Link } from 'react-router-dom';
 
-const ProfilePhoto = (): JSX.Element => {
+const EditProfile = (): JSX.Element => {
   return (
     <Box
       p={1}
-      style={{ height: '110vh', justifyContent: 'center' }}
+      style={{ height: '80vh', justifyContent: 'center' }}
       display="flex"
       justifyContent="flex-end"
       alignSelf="flex-end"
       marginRight={5}
     >
-      <h1>Profile Photo</h1>
+      <Typography style={{ fontSize: '30px', fontWeight: '100' }}>Edit Profile</Typography>
     </Box>
   );
 };
 
-export default ProfilePhoto;
+export default EditProfile;
