+++ client/src/components/Profile/ProfilePhoto.tsx
@@ -1,21 +1,20 @@
 import Button from '@mui/material/Button';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
-import { Link } from 'react-router-dom';
 
-const EditProfile = (): JSX.Element => {
+const ProfilePhoto = (): JSX.Element => {
   return (
     <Box
       p={1}
-      style={{ height: '80vh', justifyContent: 'center' }}
+      style={{ height: '110vh', justifyContent: 'center' }}
       display="flex"
       justifyContent="flex-end"
       alignSelf="flex-end"
       marginRight={5}
     >
-      <h1>Edit Profile</h1>
+      <Typography style={{ fontSize: '30px', fontWeight: '100' }}>Profile Photo</Typography>
     </Box>
   );
 };
 
-export default EditProfile;
+export default ProfilePhoto;
